requirements.txt
matplotlib==3.6.0
numpy==1.23.3
Pillow==9.2.0
Pillow==10.0.1
torch==1.12.0+cu113
torchvision==0.13.0+cu113
